Adding logic to build mod_cuwebauth for Apache 2.4
Use Docker as build environment for mod_cuwebauth so the process is
scripted and repeatable.

Add ability to reference new CUWA versions as they become availablle.

Need to tweak /usr/share/apache2/build/config_vars.mk on Ubuntu 14.04
to avoid elevation of format-security warnings to errors in calls to
ap_rprintf().  Should check with IdM on this to see if they will fix
in code.

Calls to probe apr_psprintf() are failing in the bundled configure
script.  Provided a patch to temporarily work around the issue until
more time can be spent investigating the root cause.